fix recent OW 2.0 warning W106 in dsk.c

+++ kernel/dsk.c
@@ -872,8 +872,8 @@ STATIC WORD dskerr(COUNT code)
   {
     case 1:                    /* invalid command - general failure */
       if (code & 0x08)
-        return S_ERROR | E_NOTRDY;      /* failure(E_NOTRDY); at least on yhe INT25 route,
-                                           0x8002 is returned */
+        return (WORD)((unsigned)S_ERROR | E_NOTRDY);    /* failure(E_NOTRDY); at least on yhe INT25 route,
+                                                           0x8002 is returned */
       else
         return failure(E_CMD);
